made changes to courses GUI and components related to it
made changes to the courses GUI, re-naming the class to ThirdYearCourses, added CourseButton (for constructing new buttons for courses) and CourseLevelHeader (for constructing Headers for the course level) to the helpers folder, created a seperate folder named Dashboard for dashboard and dashboard related components.
